Support entrypoint redirection
Add an example payload file which contains a nop-chain
at the end of the entry point. The nop-chain will be
used as a placeholder to patch a call instruction
to the original entrypoint from the target.

Add support in sinject for patching the x86/64 call
to the original entrypoint after the payload execution.

Refactor code and removed unused files, updated Makefile
and README.

Signed-off-by: Sebastian Ene <sebastianene@google.com>